+++ GeneralsMD/Code/GameEngine/Include/Common/GameLOD.h
@@ -170,7 +170,7 @@ public:
 	DynamicGameLODLevel getDynamicLODLevel(void) { return m_currentDynamicLOD;}
 	void init(void);	///<initialize tables of preset LOD's.
 	void setCurrentTextureReduction(Int val) {m_currentTextureReduction = val;}
-	Int getCurrentTextureReduction(void) {return m_currentTextureReduction;}
+	Int getCurrentTextureReduction(void) {return (int)m_currentTextureReduction;}
 	Int getStaticGameLODIndex(AsciiString name);
 	Int getDynamicGameLODIndex(AsciiString name);
 	inline Bool isParticleSkipped(void);
