+++ GeneralsMD/Code/GameEngine/Include/Common/PlayerTemplate.h
@@ -218,7 +218,7 @@ public:
 
 	const PlayerTemplate* getNthPlayerTemplate(Int i) const;
 	const PlayerTemplate* findPlayerTemplate(NameKeyType namekey) const;
-	inline Int getPlayerTemplateCount() const { return m_playerTemplates.size(); }
+	inline Int getPlayerTemplateCount() const { return (int)m_playerTemplates.size(); }
 	Int getTemplateNumByName(AsciiString name) const;
 	
 	// This function will fill outStringList with all the sides found in all the templates
