+++ GeneralsMD/Code/GameEngine/Include/GameLogic/Module/FlightDeckBehavior.h
@@ -139,7 +139,7 @@ public:
 	virtual Bool reserveRunway(ObjectID id, Bool forLanding);
 	virtual void releaseRunway(ObjectID id); 
 	virtual void calcPPInfo( ObjectID id, PPInfo *info );
-	virtual Int getRunwayCount() const { return m_runways.size(); }
+	virtual Int getRunwayCount() const { return (int)m_runways.size(); }
 	virtual ObjectID getRunwayReservation( Int r, RunwayReservationType type );
 	virtual void transferRunwayReservationToNextInLineForTakeoff(ObjectID id);
 	virtual Real getApproachHeight() const { return getFlightDeckBehaviorModuleData()->m_approachHeight; }
