+++ GeneralsMD/Code/GameEngine/Source/Common/RTS/AcademyStats.cpp
@@ -1081,7 +1081,7 @@ Bool AcademyStats::calculateAcademyAdvice( AcademyAdviceInfo *info )
 	info->numTips = 0;
 
 	//Build the header for each string.
-	for( Int i = 0; i < maxAdviceTips; i++ )
+	for( Int i = 0; i < (int)maxAdviceTips; i++ )
 	{
 		info->advice[ i ].format( UnicodeString( L"\n\n" ) );
 	}
