+++ GeneralsMD/Code/GameEngine/Source/Common/System/Xfer.cpp
@@ -377,7 +377,7 @@ void Xfer::xferSTLObjectIDVector( std::vector<ObjectID> *objectIDVectorData )
 	xferVersion( &version, currentVersion );
 
 	// xfer the count of the vector
-	UnsignedShort listCount = objectIDVectorData->size();
+	UnsignedShort listCount = (uint16_t)objectIDVectorData->size();
 	xferUnsignedShort( &listCount );
 	
 	// xfer vector data
@@ -444,7 +444,7 @@ void Xfer::xferSTLObjectIDList( std::list< ObjectID > *objectIDListData )
 	xferVersion( &version, currentVersion );
 
 	// xfer the count of the list
-	UnsignedShort listCount = objectIDListData->size();
+	UnsignedShort listCount = (uint16_t)objectIDListData->size();
 	xferUnsignedShort( &listCount );
 	
 	// xfer list data
@@ -510,7 +510,7 @@ void Xfer::xferSTLIntList( std::list< Int > *intListData )
 	xferVersion( &version, currentVersion );
 
 	// xfer the count of the list
-	UnsignedShort listCount = intListData->size();
+	UnsignedShort listCount = (uint16_t)intListData->size();
 	xferUnsignedShort( &listCount );
 	
 	// xfer list data
@@ -624,7 +624,7 @@ void Xfer::xferScienceVec( ScienceVec *scienceVec )
 	xferVersion( &version, currentVersion );
 
 	// count of vector
-	UnsignedShort count = scienceVec->size();
+	UnsignedShort count = (uint16_t)scienceVec->size();
 	xferUnsignedShort( &count );
 
 	if( getXferMode() == XFER_SAVE )
