+++ GeneralsMD/Code/Libraries/Source/WWVegas/WWLib/ffactory.cpp
@@ -137,7 +137,7 @@ void SimpleFileFactoryClass::Set_Sub_Directory( const char * sub_directory )
 
 void SimpleFileFactoryClass::Prepend_Sub_Directory( const char * sub_directory )
 {
-	int sub_len = strlen(sub_directory);
+	int sub_len = (int)(int)strlen(sub_directory);
 	// Overflow prevention
 	if (sub_len > 1021) {
 		WWASSERT(0);
@@ -172,7 +172,7 @@ void SimpleFileFactoryClass::Prepend_Sub_Directory( const char * sub_directory )
 
 void SimpleFileFactoryClass::Append_Sub_Directory( const char * sub_directory )
 {
-	int sub_len = strlen(sub_directory);
+	int sub_len = (int)(int)strlen(sub_directory);
 	// Overflow prevention
 	if (sub_len > 1022) {
 		WWASSERT(0);
