+++ GeneralsMD/Code/Libraries/Source/WWVegas/WWLib/trim.cpp
@@ -72,7 +72,7 @@ char* strtrim(char* buffer)
 		}
 
 		/* Clip trailing white space from the string. */
-		for (int index = strlen(buffer) - 1; index >= 0; --index) {
+		for (int index = (int)strlen(buffer) - 1; index >= 0; --index) {
 			if ((*source != 0) && ((unsigned char)buffer[index] <= 32)) {
 				buffer[index] = '\0';
 			} else {
@@ -100,7 +100,7 @@ wchar_t* wcstrim(wchar_t* buffer)
 		}
 
 		/* Clip trailing white space from the string. */
-		for (int index = wcslen(buffer) - 1; index >= 0; --index) {
+		for (int index = (int)wcslen(buffer) - 1; index >= 0; --index) {
 			if ((*source != 0) && ((unsigned int)buffer[index] <= 32)) {
 				buffer[index] = L'\0';
 			} else {
